--- a/c++/src/ColumnWriter.cc
+++ b/c++/src/ColumnWriter.cc
@@ -936,14 +936,6 @@ namespace orc {
       std::string data;
     };
 
-    struct DictEntryWithIndex {
-      DictEntryWithIndex(const char* str, size_t len, size_t index)
-          : entry(str, len), index(index) {}
-
-      DictEntry entry;
-      size_t index;
-    };
-
     SortedStringDictionary() : totalLength_(0) {
       /// Need to set empty key otherwise dense_hash_map will not work correctly
       keyToIndex_.set_empty_key(std::string_view{});
@@ -955,11 +947,8 @@ namespace orc {
     // write dictionary data & length to output buffer
     void flush(AppendOnlyBufferedStream* dataStream, RleEncoder* lengthEncoder) const;
 
-    // reorder input index buffer from insertion order to dictionary order
-    void reorder(std::vector<int64_t>& idxBuffer) const;
-
     // get dict entries in insertion order
-    void getEntriesInInsertionOrder(std::vector<const DictEntry*>&) const;
+    const std::vector<DictEntry>& getEntriesInInsertionOrder() const;
 
     // return count of entries
     size_t size() const;
@@ -971,21 +960,14 @@ namespace orc {
 
    private:
     struct LessThan {
-      bool operator()(const DictEntryWithIndex& l, const DictEntryWithIndex& r) {
-        /*
-        const auto& left = l.entry;
-        const auto& right = r.entry;
-        int ret = memcmp(left.data, right.data, std::min(left.length, right.length));
-        if (ret != 0) {
-          return ret < 0;
-        }
-        return left.length < right.length;
-        */
-        return l.entry.data < r.entry.data;  // use std::string's operator<
+      bool operator()(const DictEntry& l, const DictEntry& r) {
+        return l.data < r.data;  // use std::string's operator<
       }
     };
 
-    mutable std::vector<DictEntryWithIndex> flatDict_;
+    // store dictionary entries in insertion order
+    mutable std::vector<DictEntry> flatDict_;
+    // map from string to its insertion order index
     google::dense_hash_map<std::string_view, size_t> keyToIndex_;
     uint64_t totalLength_;
 
@@ -1005,10 +987,10 @@ namespace orc {
     if (it != keyToIndex_.end()) {
       return it->second;
     } else {
-      flatDict_.emplace_back(str, len, index);
+      flatDict_.emplace_back(str, len);
       totalLength_ += len;
 
-      const auto& lastEntry = flatDict_.back().entry;
+      const auto& lastEntry = flatDict_.back();
       keyToIndex_.emplace(std::string_view{lastEntry.data.data(), lastEntry.data.size()}, index);
       return index;
     }
@@ -1017,50 +999,16 @@ namespace orc {
   // write dictionary data & length to output buffer
   void SortedStringDictionary::flush(AppendOnlyBufferedStream* dataStream,
                                      RleEncoder* lengthEncoder) const {
-    std::sort(flatDict_.begin(), flatDict_.end(), LessThan());
-
-    for (const auto& entryWithIndex : flatDict_) {
-      const auto& entry = entryWithIndex.entry;
+    for (const auto& entry : flatDict_) {
       dataStream->write(entry.data.data(), entry.data.size());
       lengthEncoder->write(static_cast<int64_t>(entry.data.size()));
     }
   }
 
-  /**
-   * Reorder input index buffer from insertion order to dictionary order
-   *
-   * We require this function because string values are buffered by indexes
-   * in their insertion order. Until the entire dictionary is complete can
-   * we get their sorted indexes in the dictionary in that ORC specification
-   * demands dictionary should be ordered. Therefore this function transforms
-   * the indexes from insertion order to dictionary value order for final
-   * output.
-   */
-  void SortedStringDictionary::reorder(std::vector<int64_t>& idxBuffer) const {
-    // iterate the dictionary to get mapping from insertion order to value order
-    std::vector<size_t> mapping(flatDict_.size());
-    for (size_t i = 0; i < flatDict_.size(); ++i) {
-      mapping[flatDict_[i].index] = i;
-    }
-
-    // do the transformation
-    for (size_t i = 0; i != idxBuffer.size(); ++i) {
-      idxBuffer[i] = static_cast<int64_t>(mapping[static_cast<size_t>(idxBuffer[i])]);
-    }
-  }
-
   // get dict entries in insertion order
-  void SortedStringDictionary::getEntriesInInsertionOrder(
-      std::vector<const DictEntry*>& entries) const {
-    std::sort(flatDict_.begin(), flatDict_.end(),
-              [](const DictEntryWithIndex& left, const DictEntryWithIndex& right) {
-                return left.index < right.index;
-              });
-
-    entries.resize(flatDict_.size());
-    for (size_t i = 0; i < flatDict_.size(); ++i) {
-      entries[i] = &(flatDict_[i].entry);
-    }
+  const std::vector<SortedStringDictionary::DictEntry>&
+  SortedStringDictionary::getEntriesInInsertionOrder() const {
+    return flatDict_;
   }
 
   // return count of entries
@@ -1379,9 +1327,6 @@ namespace orc {
       // flush dictionary data & length streams
       dictionary.flush(dictStream.get(), dictLengthEncoder.get());
 
-      // convert index from insertion order to dictionary order
-      dictionary.reorder(dictionary.idxInDictBuffer_);
-
       // write data sequences
       int64_t* data = dictionary.idxInDictBuffer_.data();
       if (enableIndex) {
@@ -1425,16 +1370,14 @@ namespace orc {
     }
 
     // get dictionary entries in insertion order
-    std::vector<const SortedStringDictionary::DictEntry*> entries;
-    dictionary.getEntriesInInsertionOrder(entries);
+    const auto& entries = dictionary.getEntriesInInsertionOrder();
 
     // store each length of the data into a vector
-    const SortedStringDictionary::DictEntry* dictEntry = nullptr;
     for (uint64_t i = 0; i != dictionary.idxInDictBuffer_.size(); ++i) {
       // write one row data in direct encoding
-      dictEntry = entries[static_cast<size_t>(dictionary.idxInDictBuffer_[i])];
-      directDataStream->write(dictEntry->data.data(), dictEntry->data.size());
-      directLengthEncoder->write(static_cast<int64_t>(dictEntry->data.size()));
+      const auto& dictEntry = entries[static_cast<size_t>(dictionary.idxInDictBuffer_[i])];
+      directDataStream->write(dictEntry.data.data(), dictEntry.data.size());
+      directLengthEncoder->write(static_cast<int64_t>(dictEntry.data.size()));
     }
 
     deleteDictStreams();