--- a/c++/test/CMakeLists.txt
+++ b/c++/test/CMakeLists.txt
@@ -75,6 +75,7 @@ target_link_libraries (orc-test
   orc::zlib
   orc::gtest
   orc::gmock
+  orc::sparsehash
   orc-test-include
 )