--- a/c++/test/meson.build
+++ b/c++/test/meson.build
@@ -70,6 +70,7 @@ orc_test = executable(
         zlib_dep,
         gtest_dep,
         gmock_dep,
+        sparsehash_c11_dep,
     ],
 )