--- a/cmake_modules/ThirdpartyToolchain.cmake
+++ b/cmake_modules/ThirdpartyToolchain.cmake
@@ -26,13 +26,14 @@ set(ZLIB_VERSION "1.3.1")
 set(GTEST_VERSION "1.12.1")
 set(PROTOBUF_VERSION "3.5.1")
 set(ZSTD_VERSION "1.5.7")
+set(SPARSEHASH_C11_VERSION "2.11.1")
 
-option(ORC_PREFER_STATIC_PROTOBUF "Prefer static protobuf library, if available" ON)
-option(ORC_PREFER_STATIC_SNAPPY   "Prefer static snappy library, if available"   ON)
-option(ORC_PREFER_STATIC_LZ4      "Prefer static lz4 library, if available"      ON)
-option(ORC_PREFER_STATIC_ZSTD     "Prefer static zstd library, if available"     ON)
-option(ORC_PREFER_STATIC_ZLIB     "Prefer static zlib library, if available"     ON)
-option(ORC_PREFER_STATIC_GMOCK    "Prefer static gmock library, if available"    ON)
+option(ORC_PREFER_STATIC_PROTOBUF       "Prefer static protobuf library, if available"          ON)
+option(ORC_PREFER_STATIC_SNAPPY         "Prefer static snappy library, if available"            ON)
+option(ORC_PREFER_STATIC_LZ4            "Prefer static lz4 library, if available"               ON)
+option(ORC_PREFER_STATIC_ZSTD           "Prefer static zstd library, if available"              ON)
+option(ORC_PREFER_STATIC_ZLIB           "Prefer static zlib library, if available"              ON)
+option(ORC_PREFER_STATIC_GMOCK          "Prefer static gmock library, if available"             ON)
 
 # zstd requires us to add the threads
 FIND_PACKAGE(Threads REQUIRED)
@@ -99,6 +100,10 @@ endif ()
 
 if (DEFINED ENV{GTEST_HOME})
   set (GTEST_HOME "$ENV{GTEST_HOME}")
+endif ()
+
+if (DEFINED ENV{SPARSEHASH_C11_HOME})
+  set (SPARSEHASH_C11_HOME "$ENV{SPARSEHASH_C11_HOME}")
 endif ()
 
 # ----------------------------------------------------------------------
@@ -579,6 +584,42 @@ add_library (orc::protobuf ALIAS orc_protobuf)
 if (NOT (ORC_PACKAGE_KIND STREQUAL "conan" OR ORC_PACKAGE_KIND STREQUAL "vcpkg"))
   add_library (orc::protoc ALIAS orc_protoc)
 endif ()
+
+# ----------------------------------------------------------------------
+# SPARSEHASH_C11
+
+set(SPARSEHASH_C11_HOME "${THIRDPARTY_DIR}/sparsehash_c11_ep-install")
+set(SPARSEHASH_C11_INCLUDE_DIR "${SPARSEHASH_C11_HOME}/include/")
+set(SPARSEHASH_C11_STATIC_LIB_NAME sparsehash_c11)
+set(SPARSEHASH_C11_STATIC_LIB "${SPARSEHASH_C11_HOME}/lib/${SPARSEHASH_C11_STATIC_LIB_NAME}")
+set(SPARSEHASH_C11_CMAKE_ARGS
+    -DCMAKE_INSTALL_PREFIX=${SPARSEHASH_C11_HOME}
+    -DBUILD_SHARED_LIBS=OFF
+    -DCMAKE_INSTALL_LIBDIR=lib
+    -DCMAKE_POLICY_VERSION_MINIMUM=3.5
+)
+if (BUILD_POSITION_INDEPENDENT_LIB)
+  set(SPARSEHASH_C11_CMAKE_ARGS ${SPARSEHASH_C11_CMAKE_ARGS} -DCMAKE_POSITION_INDEPENDENT_CODE=ON)
+endif ()
+
+if (CMAKE_VERSION VERSION_GREATER "3.7")
+    set(SPARSEHASH_C11_CONFIGURE SOURCE_SUBDIR "" CMAKE_ARGS ${SPARSEHASH_C11_CMAKE_ARGS})
+  else()
+    set(SPARSEHASH_C11_CONFIGURE CONFIGURE_COMMAND "${THIRDPARTY_CONFIGURE_COMMAND}" ${SPARSEHASH_C11_CMAKE_ARGS}
+            "${CMAKE_CURRENT_BINARY_DIR}/sparsehash_c11_ep-prefix/src/sparsehash_c11_ep/")
+  endif()
+
+ExternalProject_Add(sparsehash_c11_ep
+    URL "https://github.com/sparsehash/sparsehash-c11/archive/refs/tags/v${SPARSEHASH_C11_VERSION}.tar.gz"
+    ${SPARSEHASH_C11_CONFIGURE}
+    ${THIRDPARTY_LOG_OPTIONS}
+    BUILD_BYPRODUCTS ${SPARSEHASH_C11_STATIC_LIB})
+
+orc_add_built_library (sparsehash_c11_ep orc_sparsehash_c11 ${SPARSEHASH_C11_STATIC_LIB} ${SPARSEHASH_C11_INCLUDE_DIR})
+list (APPEND ORC_VENDOR_DEPENDENCIES "orc::vendored_sparsehash_c11|${SPARSEHASH_C11_STATIC_LIB}")
+list (APPEND ORC_INSTALL_INTERFACE_TARGETS "$<INSTALL_INTERFACE:orc::vendored_sparsehash_c11>")
+
+add_library (orc::sparsehash_c11 ALIAS orc_sparsehash_c11)
 
 # ----------------------------------------------------------------------
 # LIBHDFSPP