@taiyang-li 2025-07-15
> Thank you for making a PR, @taiyang-li . It seems to fail on `Debian 12` with CMake. Could you fix that first?
> 
> * https://github.com/apache/orc/actions/runs/16162231691/job/45615974033?pr=2321
> 
> ```
> [==========] 744 tests from 47 test suites ran. (12344 ms total)
> [  PASSED  ] 742 tests.
> [  FAILED  ] 2 tests, listed below:
> [  FAILED  ] DictionaryEncoding.stringDictionaryWithNullWithoutIndex
> [  FAILED  ] DictionaryEncoding.multipleStripesWithIndex
> 
>  2 FAILED TESTS
> ...
> 
> The following tests FAILED:
> 	  1 - orc-test (Failed)
> make[3]: *** [CMakeFiles/test-out.dir/build.make:70: CMakeFiles/test-out] Error 8
> make[2]: *** [CMakeFiles/Makefile2:459: CMakeFiles/test-out.dir/all] Error 2
> make[1]: *** [CMakeFiles/Makefile2:466: CMakeFiles/test-out.dir/rule] Error 2
> make: *** [Makefile:303: test-out] Error 2
> FAILED debian12
> ```

Already fixed.